requirements.txt
Groq 
streamlit 
phidata
duckduckgo -search 
